+++ frontend/src/App.tsx
@@ -1,0 +1,13 @@
+import { BrowserRouter, Route, Routes } from "react-router";
+
+function App() {
+  return (
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<div>Home Route</div>} />
+      </Routes>
+    </BrowserRouter>
+  );
+}
+
+export default App;